+++ packages/extension/src/plugins/locator.ts
@@ -1,34 +1,15 @@
 import download from "download";
 import * as fs from "fs";
-import got, { CancelableRequest } from "got";
-import { flatten } from "lodash";
+import { flatten, map } from "lodash";
 import path from "path";
 import { promisify } from "util";
 import { ProgressLocation, window } from "vscode";
 import { output } from "../constants";
 import { Sysl } from "../tools/sysl";
-import {
-  CommandPluginConfig,
-  LspPluginConfig,
-  PluginConfig,
-  TransformPluginConfig,
-} from "./plugin_config";
-import { PluginClientOptions, PluginManifest, PluginManifests } from "./types";
+import { File, PluginClientOptions, PluginConfig, PluginKind, PluginManifest } from "./types";
 
 /** Name of the directory that may store Sysl metadata for a project. */
 const syslMetaDir = ".sysl";
-
-const exists = promisify(fs.exists);
-
-const idFromFile = (file: string): string => path.parse(file).name;
-
-const configFromFile = (file: string | undefined) => {
-  if (file && fs.existsSync(file)) {
-    return JSON.parse(fs.readFileSync(file).toString());
-  }
-  // If documentSelector is not defined, it may be defaulted elsewhere.
-  return {};
-};
 
 /** Discovers available plugins on the user's machine and known network locations. */
 export class PluginLocator {
@@ -44,9 +25,9 @@ export class PluginLocator {
     return flatten(
       await Promise.all([
         this.localPlugins(sysl, workspaceDirs, options),
-        this.builtin(sysl, extensionPath, options),
+        this.builtin(extensionPath, options),
         options?.workspaceConfig?.plugins?.fetchFromNetwork
-          ? this.networkPlugins(sysl, globalStoragePath, remoteUrl, options).catch((err) => {
+          ? this.networkPlugins(globalStoragePath, remoteUrl, options).catch((err) => {
               window.showErrorMessage(`Failed to fetch plugins: ${err}`);
               return [];
             })
@@ -61,67 +42,24 @@ export class PluginLocator {
     workspaceDirs: string[],
     options?: PluginClientOptions
   ): Promise<PluginConfig[]> {
-    const compatible = (filename: string): boolean => {
-      const exe = filename.endsWith(".exe");
-      return process.platform === "win32" ? exe : !exe;
-    };
-
-    async function commandPlugins(dir: string): Promise<PluginConfig[]> {
-      const scriptsPath = path.join(dir, syslMetaDir, "plugins");
-      return (await filesIn(scriptsPath)).filter(compatible).map((script) => {
-        const scriptName = path.parse(script).base;
-        return {
-          id: scriptName,
-          command: {
-            sysl,
-            runOptions: { command: script },
-            clientOptions: options,
-          },
-        } as CommandPluginConfig;
-      });
-    }
-
-    async function transformPlugins(dir: string): Promise<PluginConfig[]> {
-      const scriptsPath = path.join(dir, syslMetaDir, "diagram_renderers");
-      return (await filesIn(scriptsPath)).map(
-        (script) =>
-          ({
-            id: idFromFile(script),
-            transform: {
-              sysl,
-              scriptPath: script,
-              clientOptions: options,
-            },
-          } as TransformPluginConfig)
-      );
-    }
-
     async function lspPlugins(dir: string): Promise<PluginConfig[]> {
-      const scriptsPath = path.join(dir, syslMetaDir, "plugins", "lsp");
-      return (await dirsIn(scriptsPath)).map((dir: string) => {
-        const config = configFromFile(path.join(dir, "config.json"));
+      const scriptsPath = path.join(dir, syslMetaDir, "plugins");
+      return (await dirsIn(scriptsPath)).map((dir) => {
         return {
           id: dir,
-          lsp: {
-            scriptPath: path.join(dir, "index.js"),
-            clientOptions: { ...options, ...config },
-          },
-        } as LspPluginConfig;
+          scriptPath: path.join(dir, "index.js"),
+          clientOptions: { ...options },
+        } as PluginConfig;
       });
     }
 
     return flatten(
-      await Promise.all(
-        workspaceDirs.map(async (dir) =>
-          flatten([await commandPlugins(dir), await transformPlugins(dir), await lspPlugins(dir)])
-        )
-      )
+      await Promise.all(workspaceDirs.map(async (dir) => flatten([await lspPlugins(dir)])))
     );
   }
 
   /** Returns all plugins that are built into the extension. */
   static async builtin(
-    sysl: Sysl,
     extensionPath: string,
     options?: PluginClientOptions
   ): Promise<PluginConfig[]> {
@@ -129,187 +67,128 @@ export class PluginLocator {
     const intPath = path.join(pluginsDir, "integration", "index.js");
     const erdPath = path.join(pluginsDir, "erd", "index.js");
     const sysldPath = path.join(pluginsDir, "sysld", "index.js");
+    const clientOptions = { ...options, documentSelector: [{ scheme: "file", language: "sysl" }] };
 
     return [
-      {
-        id: "integration",
-        lsp: {
-          scriptPath: intPath,
-          clientOptions: { ...options, documentSelector: [{ scheme: "file", language: "sysl" }] },
-        },
-      } as LspPluginConfig,
-      {
-        id: "edg",
-        lsp: {
-          scriptPath: erdPath,
-          clientOptions: { ...options, documentSelector: [{ scheme: "file", language: "sysl" }] },
-        },
-      } as LspPluginConfig,
+      { id: "integration", scriptPath: intPath, clientOptions } as PluginConfig,
+      { id: "edg", scriptPath: erdPath, clientOptions } as PluginConfig,
       {
         id: "sysld",
-        lsp: {
-          scriptPath: sysldPath,
-          clientOptions: { ...options, documentSelector: [{ scheme: "file", language: "sysld" }] },
-        },
-      } as LspPluginConfig,
+        scriptPath: sysldPath,
+        clientOptions: { ...options, documentSelector: [{ scheme: "file", language: "sysld" }] },
+      } as PluginConfig,
     ];
   }
 
   /** Returns all plugins that can be discovered on the network. */
   static async networkPlugins(
-    sysl: Sysl,
     globalStoragePath: string,
     remoteUrl: string,
     options?: PluginClientOptions
   ): Promise<PluginConfig[]> {
-    return window.withProgress(
+    return window.withProgress<PluginConfig[]>(
       {
         location: ProgressLocation.Notification,
         title: "Initializing plugins",
         cancellable: true,
       },
-      async (progress, token) => {
+      async (progress, token): Promise<PluginConfig[]> => {
         progress.report({ message: "Fetching manifest...", increment: 10 });
 
-        const gotConfig = {
-          // This timeout is to pass github runner on Windows environment.
-          // FIXME: Adjust timeout if needed.
-          timeout: 10000,
-          // This can only be handled by an API behind internal DNS which is implicitly trusted.
-          https: { rejectUnauthorized: false },
-        };
-
-        return new Promise((resolve, reject) => {
-          const get = got.get(remoteUrl, gotConfig);
-          (get.json() as CancelableRequest<PluginManifests>)
-            .then(({ plugins }) => {
-              progress.report({ message: "Processing plugins...", increment: 20 });
-              this.parseManifest(sysl, globalStoragePath, remoteUrl, plugins, options).then(
-                resolve
-              );
-            })
-            .catch(reject);
-
-          token.onCancellationRequested(() => {
-            reject("Cancelled by user");
-            get?.cancel();
-          });
-        });
+        // return networkPluginsRaw(sysl, globalStoragePath, remoteUrl, progress, token, options);
+        const manifests = await resolvePlugins(globalStoragePath, remoteUrl, options);
+        return manifests.map((m) => ({
+          id: m.id,
+          scriptPath: m.entrypoint,
+          clientOptions: options,
+        }));
       }
     );
   }
-
-  /** Parses the Manifest to download plugins and/or return their configs. */
-  static async parseManifest(
-    sysl: Sysl,
-    globalStoragePath: string,
-    manifestPath: string,
-    manifests: PluginManifest[],
-    options?: PluginClientOptions
-  ): Promise<PluginConfig[]> {
-    let configs: PluginConfig[] = [];
-    try {
-      for (let i = 0; i < manifests.length; i++) {
-        const plugin = manifests[i];
-        const scriptPath = path.resolve(path.dirname(manifestPath), plugin.entrypoint);
-        switch (plugin.kind) {
-          case "command":
-            configs.push({
-              id: plugin.id,
-              command: { sysl, runOptions: { command: scriptPath }, clientOptions: options },
-            } as CommandPluginConfig);
-            break;
-          case "transform":
-            configs.push({
-              id: plugin.id,
-              transform: { sysl, scriptPath, clientOptions: options },
-            } as TransformPluginConfig);
-            break;
-          case "lsp.module":
-            const config = configFromFile(plugin.config);
-            configs.push({
-              id: plugin.id,
-              lsp: { scriptPath, clientOptions: { ...options, ...config } },
-            } as LspPluginConfig);
-            break;
-          case "archive":
-            const pluginPath: string = path.join(
-              globalStoragePath,
-              syslMetaDir,
-              "plugins",
-              plugin.id
-            );
-            // check if already downloaded (fetch from globalStorage)
-            // if ((await filesIn(pluginPath)).length) {
-            //   console.log(`plugin found at ${pluginPath}`);
-            //   // expect this plugin config to be returned by another function that discovers plugins in global storage
-            //   } else {
-            //   // if not, download the plugin and store it in globalStorage
-            //   console.log("plugin unavailable, downloading it now...");
-            const pluginManifest = await downloadPlugin(pluginPath, plugin.entrypoint);
-            // collect plugins generated from each manifest and combine them together
-            configs.push(
-              ...(await this.parseManifest(
-                sysl,
-                globalStoragePath,
-                pluginPath,
-                pluginManifest,
-                options
-              ))
-            );
-            // }
-            break;
-          default:
-            throw new Error(`Invalid plugin kind: ${plugin.kind}`);
-        }
-      }
-    } catch (e) {
-      throw new Error(`Error parsing Manifest: ${e}`);
-    } finally {
-      return configs;
-    }
-  }
 }
 
-async function downloadPlugin(dir: string, url: string): Promise<PluginManifest[]> {
-  output.appendLine(`Downloading plugin from ${url} into ${dir}...`);
+/**
+ * Returns a list of plugins rooted at the plugin defined at `url`.
+ *
+ * Each plugin is downloaded and extracted to `dir`, and `archive`-kind plugins are recursively
+ * resolved.
+ */
+export async function resolvePlugins(
+  dir: string,
+  url: string,
+  options: PluginClientOptions = {},
+  urls: string[] = []
+): Promise<PluginManifest[]> {
+  const newUrls = [...urls, url];
+  if (urls.includes(url)) {
+    throw new Error(`Cyclical dependency between plugins: ${newUrls.join(" -> ")}`);
+  }
+  const files = await downloadArchive(dir, url);
+  const plugins = extractPlugins(dir, files);
+  const children = plugins.map(
+    async (p): Promise<PluginManifest[]> =>
+      p.kind === "archive" ? resolvePlugins(dir, p.entrypoint, options, newUrls) : [p]
+  );
+  return flatten(await Promise.all(children));
+}
 
+async function downloadArchive(dir: string, url: string): Promise<File[]> {
+  output.appendLine(`Downloading plugin from ${url} into ${dir}...`);
   try {
     await promisify(fs.mkdir)(dir);
   } catch (e: any) {
-    if (e.code === "EEXIST") {
-      // Ignore error on mkdir for existing dir.
-    }
+    // Ignore error on mkdir for existing dir.
+    if (e.code !== "EEXIST") throw e;
   }
 
-  let response;
   try {
-    response = (await download(url, dir, { extract: true })) as any;
+    return (await download(url, dir, { extract: true })) as File[];
   } catch (err) {
     console.error("Error downloading plugin", err);
     throw err;
   }
-  // extract manifest.json from the archive
-  const manifest = response.find((i) => i.path === "manifest.json");
-  const plugins: PluginManifest[] = JSON.parse(manifest.data.toString())?.plugins;
-  plugins?.forEach((p: PluginManifest) => (p.entrypoint = path.resolve(dir, p.entrypoint)));
+}
+
+/** Returns an array of plugin manifests describing the plugins implemented by files. */
+export function extractPlugins(dir: string, files: File[]): PluginManifest[] {
+  const pkgJson = files.find((i) => i.path.endsWith("package.json"));
+  if (!pkgJson) throw new Error(`no package plugins described by ${files.length} files`);
+
+  const plugins: PluginManifest[] = [];
+  const pkg = JSON.parse(pkgJson.data.toString());
+  // `main` indicates an entrypoint to an LSP plugin.
+  if (pkg.main) {
+    plugins.push({
+      id: pkg.name,
+      version: pkg.version,
+      entrypoint: path.resolve(dir, `package/${pkg.main}`),
+      kind: "lsp.module",
+    });
+  }
+
+  // `peerDependencies` reference child plugins that should be resolved.
+  if (pkg.peerDependencies) {
+    plugins.push(
+      ...map(pkg.peerDependencies, (version, name) => {
+        const uriName = encodeURIComponent(name);
+        const registry = (pkg.publishConfig.registry as string).replace(/\/$/, "");
+        return {
+          id: name,
+          version,
+          kind: "archive" as PluginKind,
+          entrypoint: [registry, uriName, "-", `${uriName}-${version}.tgz`].join("/"),
+        };
+      })
+    );
+  }
   return plugins;
 }
 
 async function dirsIn(dir: string): Promise<string[]> {
-  if (!(await exists(dir))) {
+  if (!(await promisify(fs.exists)(dir))) {
     return [];
   }
   return (await promisify(fs.readdir)(dir, { withFileTypes: true }))
     .filter((i) => i.isDirectory())
-    .map((i) => path.join(dir, i.name));
-}
-/** Returns the absolute path to each file in dir. */
-async function filesIn(dir: string): Promise<string[]> {
-  if (!(await exists(dir))) {
-    return [];
-  }
-  return (await promisify(fs.readdir)(dir, { withFileTypes: true }))
-    .filter((i) => !i.isDirectory())
     .map((i) => path.join(dir, i.name));
 }